feat: split test models to statetest-tests
