+++ bins/revme/src/cmd/statetest.rs
@@ -1,5 +1,4 @@
 pub mod merkle_trie;
-pub mod models;
 mod runner;
 pub mod utils;
