--- bins/revme/src/cmd/statetest/models/mod.rs
@@ -1,198 +1,0 @@
-mod deserializer;
-mod spec;
-
-use deserializer::*;
-
-pub use spec::SpecName;
-
-use revm::{
-    primitives::{Address, Bytes, HashMap, B256, U256},
-    specification::{
-        eip2930::AccessList,
-        eip7702::{Authorization, Parity, RecoveredAuthorization, Signature},
-    },
-};
-use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
-
-#[derive(Debug, PartialEq, Eq, Deserialize)]
-pub struct TestSuite(pub BTreeMap<String, TestUnit>);
-
-#[derive(Debug, PartialEq, Eq, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct TestUnit {
-    /// Test info is optional
-    #[serde(default, rename = "_info")]
-    pub info: Option<serde_json::Value>,
-
-    pub env: Env,
-    pub pre: HashMap<Address, AccountInfo>,
-    pub post: BTreeMap<SpecName, Vec<Test>>,
-    pub transaction: TransactionParts,
-    #[serde(default)]
-    pub out: Option<Bytes>,
-}
-
-/// State test indexed state result deserialization.
-#[derive(Debug, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct Test {
-    pub expect_exception: Option<String>,
-
-    /// Indexes
-    pub indexes: TxPartIndices,
-    /// Post state hash
-    pub hash: B256,
-    /// Post state
-    #[serde(default)]
-    pub post_state: HashMap<Address, AccountInfo>,
-
-    /// Logs root
-    pub logs: B256,
-
-    /// Tx bytes
-    pub txbytes: Option<Bytes>,
-}
-
-#[derive(Debug, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct TxPartIndices {
-    pub data: usize,
-    pub gas: usize,
-    pub value: usize,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct AccountInfo {
-    pub balance: U256,
-    pub code: Bytes,
-    #[serde(deserialize_with = "deserialize_str_as_u64")]
-    pub nonce: u64,
-    pub storage: HashMap<U256, U256>,
-}
-
-#[derive(Debug, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct Env {
-    pub current_coinbase: Address,
-    #[serde(default)]
-    pub current_difficulty: U256,
-    pub current_gas_limit: U256,
-    pub current_number: U256,
-    pub current_timestamp: U256,
-    pub current_base_fee: Option<U256>,
-    pub previous_hash: Option<B256>,
-
-    pub current_random: Option<B256>,
-    pub current_beacon_root: Option<B256>,
-    pub current_withdrawals_root: Option<B256>,
-
-    pub parent_blob_gas_used: Option<U256>,
-    pub parent_excess_blob_gas: Option<U256>,
-    pub current_excess_blob_gas: Option<U256>,
-}
-
-#[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TransactionParts {
-    pub data: Vec<Bytes>,
-    pub gas_limit: Vec<U256>,
-    pub gas_price: Option<U256>,
-    pub nonce: U256,
-    pub secret_key: B256,
-    /// if sender is not present we need to derive it from secret key.
-    #[serde(default)]
-    pub sender: Option<Address>,
-    #[serde(default, deserialize_with = "deserialize_maybe_empty")]
-    pub to: Option<Address>,
-    pub value: Vec<U256>,
-    pub max_fee_per_gas: Option<U256>,
-    pub max_priority_fee_per_gas: Option<U256>,
-
-    #[serde(default)]
-    pub access_lists: Vec<Option<AccessList>>,
-    pub authorization_list: Option<Vec<TestAuthorization>>,
-    #[serde(default)]
-    pub blob_versioned_hashes: Vec<B256>,
-    pub max_fee_per_blob_gas: Option<U256>,
-}
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, Copy)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct TestAuthorization {
-    chain_id: U256,
-    address: Address,
-    nonce: U256,
-    v: U256,
-    r: U256,
-    s: U256,
-    signer: Option<Address>,
-}
-
-impl TestAuthorization {
-    pub fn signature(&self) -> Signature {
-        let v = u64::try_from(self.v).unwrap_or(u64::MAX);
-        let parity = Parity::try_from(v).unwrap_or(Parity::Eip155(36));
-        Signature::from_rs_and_parity(self.r, self.s, parity).unwrap()
-    }
-
-    pub fn into_recovered(self) -> RecoveredAuthorization {
-        let authorization = Authorization {
-            chain_id: self.chain_id,
-            address: self.address,
-            nonce: u64::try_from(self.nonce).unwrap(),
-        };
-        let authority = self
-            .signature()
-            .recover_address_from_prehash(&authorization.signature_hash())
-            .ok();
-        RecoveredAuthorization::new_unchecked(
-            authorization.into_signed(self.signature()),
-            authority,
-        )
-    }
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-    use serde_json::Error;
-
-    #[test]
-    pub fn serialize_u256() -> Result<(), Error> {
-        let json = r#"{"_item":"0x10"}"#;
-
-        #[derive(Deserialize, Debug)]
-        pub struct Test {
-            _item: Option<U256>,
-        }
-
-        let out: Test = serde_json::from_str(json)?;
-        println!("out:{out:?}");
-        Ok(())
-    }
-
-    #[test]
-    pub fn deserialize_minimal_transaction_parts() -> Result<(), Error> {
-        let json = r#"{"data":[],"gasLimit":[],"nonce":"0x0","secretKey":"0x0000000000000000000000000000000000000000000000000000000000000000","to":"","value":[]}"#;
-
-        let _: TransactionParts = serde_json::from_str(json)?;
-        Ok(())
-    }
-
-    #[test]
-    pub fn serialize_b160() -> Result<(), Error> {
-        let json = r#"{"_item":"0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba"}"#;
-
-        #[derive(Deserialize, Debug)]
-        pub struct Test {
-            _item: Address,
-        }
-
-        let out: Test = serde_json::from_str(json)?;
-        println!("out:{out:?}");
-        Ok(())
-    }
-}
