+++ crates/statetest-types/src/env.rs
@@ -1,0 +1,23 @@
+use revm::primitives::{Address, B256, U256};
+use serde::Deserialize;
+
+#[derive(Debug, PartialEq, Eq, Deserialize)]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
+pub struct Env {
+    pub current_coinbase: Address,
+    #[serde(default)]
+    pub current_difficulty: U256,
+    pub current_gas_limit: U256,
+    pub current_number: U256,
+    pub current_timestamp: U256,
+    pub current_base_fee: Option<U256>,
+    pub previous_hash: Option<B256>,
+
+    pub current_random: Option<B256>,
+    pub current_beacon_root: Option<B256>,
+    pub current_withdrawals_root: Option<B256>,
+
+    pub parent_blob_gas_used: Option<U256>,
+    pub parent_excess_blob_gas: Option<U256>,
+    pub current_excess_blob_gas: Option<U256>,
+}
