+++ crates/statetest-types/src/transaction.rs
@@ -1,0 +1,40 @@
+use revm::{
+    primitives::{Address, Bytes, B256, U256},
+    specification::eip2930::AccessList,
+};
+use serde::{Deserialize, Serialize};
+
+use crate::{deserializer::deserialize_maybe_empty, TestAuthorization};
+
+#[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct TransactionParts {
+    pub data: Vec<Bytes>,
+    pub gas_limit: Vec<U256>,
+    pub gas_price: Option<U256>,
+    pub nonce: U256,
+    pub secret_key: B256,
+    /// if sender is not present we need to derive it from secret key.
+    #[serde(default)]
+    pub sender: Option<Address>,
+    #[serde(default, deserialize_with = "deserialize_maybe_empty")]
+    pub to: Option<Address>,
+    pub value: Vec<U256>,
+    pub max_fee_per_gas: Option<U256>,
+    pub max_priority_fee_per_gas: Option<U256>,
+
+    #[serde(default)]
+    pub access_lists: Vec<Option<AccessList>>,
+    pub authorization_list: Option<Vec<TestAuthorization>>,
+    #[serde(default)]
+    pub blob_versioned_hashes: Vec<B256>,
+    pub max_fee_per_blob_gas: Option<U256>,
+}
+
+#[derive(Debug, PartialEq, Eq, Deserialize)]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
+pub struct TxPartIndices {
+    pub data: usize,
+    pub gas: usize,
+    pub value: usize,
+}
